refactor: refactored evaluation code
