+++ .gitignore
@@ -58,5 +58,5 @@ noise2void
 
 utils/old
 schemas/phf
-
+*/old/
 configs
