+++ src/ssm/schemas/__init__.py
@@ -1,0 +1,1 @@
+from .baselines import *
