+++ src/ssm/schemas/baselines/__init__.py
@@ -1,0 +1,1 @@
+from .bm3d import *
